refactor: 매직넘버 상수화
