+++ src/main/java/webserver/request/RequestHeader.java
@@ -1,5 +1,7 @@
 package webserver.request;
 
+import webserver.HttpHeader;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.HashMap;
