+++ src/main/java/webserver/response/Response.java
@@ -1,6 +1,7 @@
 package webserver.response;
 
 import webserver.FileType;
+import webserver.HttpHeader;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -8,18 +9,19 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 public class Response {
+
     private final StatusCode statusCode;
-    private final Map<String, String> responseHeader;
+    private final Map<HttpHeader, String> responseHeader;
     private final byte[] body;
 
-    private Response(StatusCode statusCode, Map<String, String> responseHeader, byte[] body) {
+    private Response(StatusCode statusCode, Map<HttpHeader, String> responseHeader, byte[] body) {
         this.statusCode = statusCode;
         this.responseHeader = responseHeader;
         this.body = body;
     }
 
     public static Response ok(byte[] body, FileType fileType) {
-        Map<String, String> responseHeader = generateResponseHeader(body, fileType);
+        Map<HttpHeader, String> responseHeader = generateResponseHeader(body, fileType);
         return new Response(
                 StatusCode.OK,
                 responseHeader,
@@ -28,8 +30,8 @@ public class Response {
     }
 
     public static Response found(byte[] body, FileType fileType, String location) {
-        Map<String, String> responseHeader = generateResponseHeader(body, fileType);
-        responseHeader.put("Location", location);
+        Map<HttpHeader, String> responseHeader = generateResponseHeader(body, fileType);
+        responseHeader.put(HttpHeader.LOCATION, location);
         return new Response(
                 StatusCode.FOUND,
                 responseHeader,
@@ -37,19 +39,19 @@ public class Response {
         );
     }
 
-    private static Map<String, String> generateResponseHeader(byte[] body, FileType fileType) {
-        Map<String, String> responseHeader = new LinkedHashMap<>();
-        responseHeader.put("Content-Type", fileType.getContentType() + ";charset=utf-8");
+    private static Map<HttpHeader, String> generateResponseHeader(byte[] body, FileType fileType) {
+        Map<HttpHeader, String> responseHeader = new LinkedHashMap<>();
+        responseHeader.put(HttpHeader.CONTENT_TYPE, fileType.getContentType() + ";charset=utf-8");
         if (body.length > 0) {
-            responseHeader.put("Content-Length", String.valueOf(body.length));
+            responseHeader.put(HttpHeader.CONTENT_LENGTH, String.valueOf(body.length));
         }
         return responseHeader;
     }
 
     public void flush(DataOutputStream dos) throws IOException {
         dos.writeBytes("HTTP/1.1 " + statusCode.getCode() + " " + statusCode.getMessage() + " \r\n");
-        for (Map.Entry<String, String> entry : responseHeader.entrySet()) {
-            dos.writeBytes(entry.getKey() + ": " + entry.getValue() + " \r\n");
+        for (Map.Entry<HttpHeader, String> header : responseHeader.entrySet()) {
+            dos.writeBytes(header.getKey().getHeaderValue() + ": " + header.getValue() + " \r\n");
         }
         dos.writeBytes("\r\n");
         dos.write(body, 0, body.length);
